+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## In development
 
 ### Added
+* Added exclude entity parama in search_entries_for_simple.
+  Contributed by @hinashi
+* Added exclude entity parama in get_referred_objects.
+  Contributed by @hinashi
 
 ### Changed
 * Changed to allow parallel execution some job.